+++ ng_trajectory/plot.py
@@ -1,6 +1,107 @@
 #!/usr/bin/env python3.6
 # plot.py
-"""Plot functions for ng_trajectory.
+"""## Plot functions for ng_trajectory
+
+From the user side (i.e., configuration file), only dynamic plotting is available.
+However, all plotting (package-wise) should be controlled by variable `plot` that
+is set to False by default.
+
+Dynamic plotting is defined using a custom key in the JSON. Following example
+resembles what is a "standard" and most used configuration:
+```json
+"plot": true,
+"plot_args": [
+    {
+        "_figure": {
+            "function": "axis",
+            "_args": [ "equal" ]
+        },
+        "trackPlot": [ "@track" ]
+    },
+    {
+        "pointsPlot": {
+            "_args": [ "@result" ]
+        },
+        "pointsScatter": {
+            "_args": [ "@rcandidate" ]
+        }
+    }
+]
+```
+Note: This creates a figure with equal axis, underlying track, optimized control
+points of the trajectory and their interpolation.
+
+The list in `plot_args` contains two dictionaries. The first one is executed
+before the optimization (and even before initialization of algorithms), whereas
+the second one is executed after optimization finishes.
+
+Commands in the `plot_args` are executed in order, key-wise. Basic syntax is
+```json
+"func": [ "arg1", "arg2" ]
+```
+
+which sends all arguments to the function, or
+```json
+"func": {
+    "_args": [ "arg1", "arg2" ],
+    "kw_arg": 4
+}
+```
+
+which calls `func` with the arguments stored in `_args`, appended with other
+arguments as kwargs.
+
+Note: Keys starting with `_` are treated differently; and are removed from kwargs.
+
+Since it is not possible to have a dictionary with repeating keys, you can use
+a meta character `-`. Dash, and everything after it is discarded during dynamic
+plotting.
+
+To pass a variable to the function, write its name prefixed with `@`. In case
+that the variable is not available, an exception is raised.
+
+
+### Available functions
+Following functions are available for plotting, however only the first three
+are usually used:
+
+- trackPlot
+- pointsScatter
+- pointsPlot
+- bordersPlot
+- indicesPlot
+- groupsScatter
+- groupsPlot
+- grouplayersScatter
+- grouplayersPlot
+- labelText
+
+
+### Available variables
+Following variables are available for plotting (by setting the value to `@` + name
+of the variable):
+
+- track -- All valid points of the track.
+- fitness -- Fitness value of the best solution.
+- rcandidate -- Control points of the best solution.
+- tcandidate -- Control points of the best solution in Matryoshka space.
+- result -- Optimized trajectory (interpolation of rcandidate).
+- figure -- Currently used figure for plotting.
+- + any variable defined in the current loop from the configuration file.
+
+
+### Matplotlib wrapper
+In addition, it is possible to call literally any function related to `pyplot` or
+`figure` from the matplotlib. To do this, call function `_pyplot`/`_figure`
+with argument `function` with the name of the required function.
+
+For example, to make the axis equal, one can use this:
+```json
+"_figure": {
+    "function": "axis",
+    "_args": [ "equal" ]
+}
+```
 """
 ######################
 # Imports & Globals
@@ -11,12 +112,18 @@ import numpy, sys
 
 try:
     from ng_trajectory import matplotlib, pyplot
+    import matplotlib._pylab_helpers
 except:
     pass
 
 from ng_trajectory import PLOT_AVAILABLE
 
 from typing import List, Dict
+
+
+# Global variables
+CANVAS = None
+CANVAS_OLD = None
 
 
 ######################
@@ -47,8 +154,36 @@ def figureCreate() -> matplotlib.figure.Figure:
     Returns:
     figure -- created Figure
     """
+    global CANVAS, CANVAS_OLD
+
     figure = pyplot.figure(dpi=300)
     figure.add_subplot(111)
+
+    try:
+        # Hold onto the PhotoImage object (canvas) in the memory to avoid gc related errors:
+        # Exception ignored in: <bound method Image.__del__ of <tkinter.PhotoImage object at 0x7f77550ab240>>
+        # Traceback (most recent call last):
+        #   File "/usr/lib/python3.6/tkinter/__init__.py", line 3519, in __del__
+        #     self.tk.call('image', 'delete', self.name)
+        # RuntimeError: main thread is not in main loop
+        CANVAS_OLD = CANVAS
+
+        # This is not very investigated, but basically any thread can run garbage collection
+        # of any variable, even tkinter of another thread.
+        # https://stackoverflow.com/questions/44781806/tkinter-objects-being-garbage-collected-from-the-wrong-thread
+        # https://bugs.python.org/issue39093
+
+        # Invoking garbage collection here (or during 'figureClose') does not work.
+        # I assume that some part of the canvas is still kept in the memory, that said
+        # calling 'delete' on tk image does not actually delete it right away.
+
+        # Should be same as
+        #    matplotlib._pylab_helpers.Gcf.get_fig_manager(1).canvas._tkphoto
+        # or matplotlib._pylab_helpers.Gcf.get_all_fig_managers()[0].canvas._tkphoto
+        CANVAS = matplotlib._pylab_helpers.Gcf.get_active().canvas._tkphoto
+    except:
+        pass
+
     return figure